requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
python-dotenv>=0.19.0
requests>=2.25.0
streamlit>=1.0.0